Setup StackNavigator for screen switching, refactored code
